server/requirements.txt
aiofiles==23.2.1
asgiref==3.7.2
click==8.1.7
fastapi==0.68.2
greenlet==3.0.3
h11==0.14.0
mysqlclient==2.2.3
pydantic==1.10.14
python-dotenv==1.0.1
SQLAlchemy==2.0.25
starlette==0.14.2
typing_extensions==4.9.0
uvicorn==0.15.0
